+++ scp/7_FP_ex.cpp
@@ -1,4 +1,4 @@
-//copyright by ArthurZhou @ UMich&FUDAN&HUST
+//copyright by ArthurZhou @ UMich&Fudan&HUST
 #include <stdlib.h>
 #include <iostream>
 #include <string>
@@ -231,6 +231,7 @@ int fp_ex(string WD_dir, string fasta, string chr, string t, int tsd_index){
      
             
     //ref pull out for 5mer
+            /*
             int start, end;
             start=loc[i][4]-loc[i][6];
             if(start<=0){
@@ -252,7 +253,7 @@ int fp_ex(string WD_dir, string fasta, string chr, string t, int tsd_index){
             ss_e.clear();
             ss_e<<end;
             s_end=ss_e.str();
-            
+            */
             //No need ref.fasta for now
             /*
             ofstream file20;
@@ -272,7 +273,7 @@ int fp_ex(string WD_dir, string fasta, string chr, string t, int tsd_index){
             system(syst_ref);
             //getchar();
             */
-            
+//Four -fold or seven -fold
     //ref pull out for fake junc
             int start_junc, end_junc;
             start_junc=loc[i][4]-J_BIN*7;
@@ -469,7 +470,7 @@ int fp_ex(string WD_dir, string fasta, string chr, string t, int tsd_index){
                     string fasta5_str_length=to_string(fasta5_str_len);
                     string fasta3_str_length=to_string(fasta3_str_len);
                     
-                    string sys_3_blast = "bash -c \"blastn -evalue 0.05 -task blastn -query <(echo -e \\\">"+name_tag_1+"\\\\n"+(fasta5_str)+"\\\") -subject "+ref_junc_file+" -dust no -outfmt \\\"7 std\\\" |grep -v \\\"#\\\" | awk '{if(\\\$3>=80&&\\\$4>="+fasta5_str_length+"&&(\\\$10-\\\$9)>0) print \\\"1\\\"}' |wc -l \"";
+                    string sys_3_blast = "bash -c \"blastn -evalue 0.05 -task blastn -query <(printf "+(fasta5_str)+") -subject "+ref_junc_file+" -dust no -outfmt \\\"7 std\\\" |grep -v \\\"#\\\" | awk '{if(\\\$3>=80&&\\\$4>="+fasta5_str_length+"&&(\\\$10-\\\$9)>0) print \\\"1\\\"}' |wc -l \"";
                     //> "+WD_dir+"read_result_junc_fake.txt";
                     //cout<<sys_3_blast<<endl;
                     
@@ -491,7 +492,7 @@ int fp_ex(string WD_dir, string fasta, string chr, string t, int tsd_index){
                     //loc_tsd_fp[w][0]=pp_3_int;
                     loc_tsd_fp[w][0]=accumulate(pp_3_int.begin(),pp_3_int.end(),loc_tsd_fp[w][0]);
                     
-                    string sys_3_blast_2 = "bash -c \"blastn -evalue 0.05 -task blastn -query <(echo -e \\\">"+name_tag_2+"\\\\n"+(fasta3_str)+"\\\") -subject "+ref_junc_file+" -dust no -outfmt \\\"7 std\\\" |grep -v \\\"#\\\" | awk '{if(\\\$3>=80&&\\\$4>="+fasta3_str_length+"&&(\\\$10-\\\$9)>0) print \\\"1\\\"}' |wc -l \"";
+                    string sys_3_blast_2 = "bash -c \"blastn -evalue 0.05 -task blastn -query <(printf "+(fasta3_str)+") -subject "+ref_junc_file+" -dust no -outfmt \\\"7 std\\\" |grep -v \\\"#\\\" | awk '{if(\\\$3>=80&&\\\$4>="+fasta3_str_length+"&&(\\\$10-\\\$9)>0) print \\\"1\\\"}' |wc -l \"";
                     //> "+WD_dir+"read_result_junc_fake.txt";
                     //cout<<sys_3_blast_2<<endl;
